// src/types/Enums.ts
// AN ENUM FOR THE TODOCONTEXTREDUCERACTION TYPE
export enum TODO_CONTEXT_REDUCER_ACTION_TYPE{
    CREATE_TODO = "CREATE_TODO",
    ADD_TODO = "ADD_TODO",
    TOGGLE_TODO = "TOGGLE_TODO",
    REMOVE_TODO = "REMOVE_TODO",
    MARK_COMPLETE = "MARK_COMPLETE",
    MARK_INCOMPLETE = "MARK_INCOMPLETE",
    MARK_ALL_COMPLETE = "MARK_ALL_COMPLETE",
    MARK_ALL_INCOMPLETE = "MARK_ALL_INCOMPLETE",
}